Use ::Project namespace for LifeCycle, Stage and Gate
