+++ app/models/work_package.rb
@@ -54,7 +54,7 @@ class WorkPackage < ApplicationRecord
   belongs_to :assigned_to, class_name: "Principal", optional: true
   belongs_to :responsible, class_name: "Principal", optional: true
   belongs_to :version, optional: true
-  belongs_to :project_life_cycle, optional: true
+  belongs_to :project_life_cycle, class_name: "Project::LifeCycle", optional: true
   belongs_to :priority, class_name: "IssuePriority"
   belongs_to :category, class_name: "Category", optional: true
