+++ spec/models/life_cycle_spec.rb
