+++ spec/models/project/gate_spec.rb
@@ -29,23 +29,29 @@
 require "rails_helper"
 require "support/shared/project_life_cycle_helpers"
 
-RSpec.describe Projects::Stage do
-  it_behaves_like "a ProjectLifeCycle event"
+RSpec.describe Project::Gate do
+  it_behaves_like "a Project::LifeCycle event"
 
   describe "validations" do
-    it { is_expected.to validate_presence_of(:start_date) }
-    it { is_expected.to validate_presence_of(:end_date) }
+    it { is_expected.to validate_presence_of(:date) }
 
-    it "is valid when `start_date` and `end_date` are present" do
-      valid_stage = build(:project_stage)
-      expect(valid_stage).to be_valid
+    it "is invalid if `start_date` is present" do
+      gate_with_start_date = build(:project_gate, start_date: Time.zone.today)
+
+      expect(gate_with_start_date).not_to be_valid
+      expect(gate_with_start_date.errors[:base]).to include("Cannot assign `start_date` or `end_date` to a Gate")
     end
 
-    it "adds an error when `date` is set" do
-      stage_with_date = build(:project_stage, date: Time.zone.today)
+    it "is invalid if `end_date` is present" do
+      gate_with_end_date = build(:project_gate, end_date: Time.zone.today)
 
-      expect(stage_with_date).not_to be_valid
-      expect(stage_with_date.errors[:base]).to include("Cannot assign `date` to a Stage")
+      expect(gate_with_end_date).not_to be_valid
+      expect(gate_with_end_date.errors[:base]).to include("Cannot assign `start_date` or `end_date` to a Gate")
+    end
+
+    it "is valid if neither `start_date` nor `end_date` are present" do
+      valid_gate = build(:project_gate)
+      expect(valid_gate).to be_valid
     end
   end
 end
